--- a/src/components/checkbox/types.ts
+++ b/src/components/checkbox/types.ts
@@ -1,9 +1,0 @@
-export type CheckboxParts =
-  | "container"
-  | "input"
-  | "option"
-  | "label"
-  | "checked"
-  | "disabled"
-  | "indeterminate"
-  | "unchecked";