--- a/src/components/tree-view/internal/tree-view-item/tree-view-item.tsx
+++ b/src/components/tree-view/internal/tree-view-item/tree-view-item.tsx
@@ -1079,9 +1079,9 @@ export class ChTreeViewItem {
 
           {(showAllLines || showLastLine) && (
             <div
-              key="dashed-line"
+              key="line"
               class={{
-                "dashed-line": true,
+                line: true,
                 "last-all-line": showAllLines && this.lastItem,
                 "last-line": showLastLine
               }}