--- a/src/components/tree-view/internal/tree-view/tree-view.tsx
+++ b/src/components/tree-view/internal/tree-view/tree-view.tsx
@@ -92,7 +92,7 @@ const POSITION_Y_DRAG_CUSTOM_VAR = "--ch-tree-view-dragging-item-y";
 let autoId = 0;
 
 /**
- * @part drag-preview - The element that contains the information to preview the current drag.
+ * @part drag-preview - The element that contains the preview information for the current drag.
  */
 @Component({
   tag: "ch-tree-view",