**Changes we propose in this PR**:
 - Fixes:
   - Fixed drag and drop precision. In some cases the leafs were taken into account to process the drop operations.

   - Update `ch-checkbox` to use the `reserved-names.ts` file for part names.

 - Features:
   - Improve "tokenMap" utility performance.

   - Add a global MutationObserver to subscribe and observe RTL changes.

   - Added support for Shadow DOM in `ch-tree-view` and `ch-tree-view-render` controls.

     This change improves the scripting time by a reduction of `33%` - `40%`. In a Tree View with 4000 items on the initial load, scripting time was reduced from `450ms` - `500ms` to `300ms`. With a 4x slowdown, scripting time was reduced from `2400ms` - `2600ms` to `1600ms`.

   - Improved Tree View documentation.

## Breaking changes
 - This PR refactors all part names used in the Tree View. Therefore, please refer to the updated readme file for all new part names.

 - Additionally, some custom var names and their values were adjusted:
   - ```scss
     Before: --ch-tree-view-item__dashed-line--inset-inline-start: 0px
     Now:    --ch-tree-view-item__line--inset-inline-start: 0px
     ```

   - ```scss
     Before: --ch-tree-view-item__dashed-line--inset-inline-end: 4px
     Now:    --ch-tree-view-item__line--inset-inline-end: 0px
     ``` 

   - ```scss
     Before: --ch-tree-view-item-gap: 5px
     Now:    --ch-tree-view-item-gap: 0px
     ``` 
 - Removed some default styles to improve white-label support. Removed:
   - `min-block-size` in the item's header
   - Disabled style in items
   - default `padding-block` in the item's action element.

 - Classes are no longer used to style the Tree View (due to shadow boundary).

 - The `class` property no longer applies to items. Now there is a new property `parts` that comes in replacement.